% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UN_plot.R
\name{random_normal}
\alias{random_normal}
\title{normal number generator}
\usage{
random_normal(n)
}
\arguments{
\item{n}{the number of values}
}
\value{
a dataframe of values and a histogram
}
\description{
generate some number of values from normal distribution and their corresponding histogram
}
\examples{
random_normal(100)
}
