% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UN_plot.R
\name{random_unif}
\alias{random_unif}
\title{uniform number generator}
\usage{
random_unif(n)
}
\arguments{
\item{n}{the number of values}
}
\value{
a dataframe of values and a histogram
}
\description{
generate some number of values from uniform distribution and their corresponding histogram
}
\examples{
random_unif(100)
}
